Quick fix: Moves the assert checking for mpi4py not in sys.modules to the add_option() function of pytest
This allows to to avoid crashing in some instances for users whose environnement adds mpi4py unexpectedly since the add_option() function that parses for the arguments of the program seems to be executed before python loads its modules. This should be tested further to see if it works for all users.